+++ DaemonMC/Network/Bedrock/RequestNetworkSettings.cs
@@ -7,7 +7,7 @@
 
     public class RequestNetworkSettings
     {
-        public const byte id = 193;
+        public const int id = 193;
         public static void Decode(byte[] buffer)
         {
             var packet = new RequestNetworkSettingsPacket
